+++ Services/PdfExporter.cs
@@ -23,7 +23,7 @@ public class IdevsPdfExporter : IIdevsPdfExporter
         string headerTemplate = "<p></p>",
         string footerTemplate = "<p></p>")
     {
-        using var browserFetcher = new BrowserFetcher();
+        var browserFetcher = new BrowserFetcher();
         await browserFetcher.DownloadAsync();
         await using var browser = await Puppeteer.LaunchAsync(new LaunchOptions { Headless = true });
         await using var page = await browser.NewPageAsync();